backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.8
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
